requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
pyswip
python-multipart